+++ solc/CommandLineInterface.cpp
@@ -170,7 +170,8 @@ void CommandLineInterface::handleEVMAssembly(std::string const& _contract)
 {
 	solAssert(m_assemblyStack);
 	solAssert(
-		CompilerInputModes.count(m_options.input.mode) == 1
+		CompilerInputModes.count(m_options.input.mode) == 1 ||
+		m_options.input.mode == frontend::InputMode::EVMAssemblerJSON
 	);
 
 	if (!m_options.compiler.outputs.asm_ && !m_options.compiler.outputs.asmJson)
@@ -194,8 +195,11 @@ void CommandLineInterface::handleEVMAssembly(std::string const& _contract)
 
 void CommandLineInterface::handleBinary(std::string const& _contract)
 {
-	solAssert(CompilerInputModes.count(m_options.input.mode) == 1);
 	solAssert(m_assemblyStack);
+	solAssert(
+		CompilerInputModes.count(m_options.input.mode) == 1 ||
+		m_options.input.mode == frontend::InputMode::EVMAssemblerJSON
+	);
 
 	std::string binary;
 	std::string binaryRuntime;
@@ -228,8 +232,11 @@ void CommandLineInterface::handleBinary(std::string const& _contract)
 
 void CommandLineInterface::handleOpcode(std::string const& _contract)
 {
-	solAssert(CompilerInputModes.count(m_options.input.mode) == 1);
 	solAssert(m_assemblyStack);
+	solAssert(
+		CompilerInputModes.count(m_options.input.mode) == 1 ||
+		m_options.input.mode == frontend::InputMode::EVMAssemblerJSON
+	);
 
 	std::string opcodes{evmasm::disassemble(m_assemblyStack->object(_contract).bytecode, m_options.output.evmVersion)};
 
@@ -330,7 +337,10 @@ void CommandLineInterface::handleIROptimizedAst(std::string const& _contractName
 
 void CommandLineInterface::handleBytecode(std::string const& _contract)
 {
-	solAssert(CompilerInputModes.count(m_options.input.mode) == 1);
+	solAssert(
+		CompilerInputModes.count(m_options.input.mode) == 1 ||
+		m_options.input.mode == frontend::InputMode::EVMAssemblerJSON
+	);
 
 	if (m_options.compiler.outputs.opcodes)
 		handleOpcode(_contract);
@@ -759,6 +769,12 @@ void CommandLineInterface::processInput()
 		compile();
 		outputCompilationResults();
 		break;
+	case InputMode::EVMAssemblerJSON:
+		assembleFromEVMAssemblyJSON();
+		handleCombinedJSON();
+		handleBytecode(m_assemblyStack->contractNames().front());
+		handleEVMAssembly(m_assemblyStack->contractNames().front());
+		break;
 	}
 }
 
@@ -775,11 +791,38 @@ void CommandLineInterface::printLicense()
 	sout() << licenseText << std::endl;
 }
 
+void CommandLineInterface::assembleFromEVMAssemblyJSON()
+{
+	solAssert(m_options.input.mode == InputMode::EVMAssemblerJSON);
+	solAssert(!m_assemblyStack);
+	solAssert(!m_evmAssemblyStack && !m_compiler);
+
+	solAssert(m_fileReader.sourceUnits().size() == 1);
+	auto&& [sourceUnitName, source] = *m_fileReader.sourceUnits().begin();
+
+	auto evmAssemblyStack = std::make_unique<evmasm::EVMAssemblyStack>(m_options.output.evmVersion);
+	try
+	{
+		evmAssemblyStack->parseAndAnalyze(sourceUnitName, source);
+	}
+	catch (evmasm::AssemblyImportException const& _exception)
+	{
+		solThrow(CommandLineExecutionError, "Assembly Import Error: "s + _exception.what());
+	}
+
+	if (m_options.output.debugInfoSelection.has_value())
+		evmAssemblyStack->selectDebugInfo(m_options.output.debugInfoSelection.value());
+	evmAssemblyStack->assemble();
+
+	m_evmAssemblyStack = std::move(evmAssemblyStack);
+	m_assemblyStack = m_evmAssemblyStack.get();
+}
+
 void CommandLineInterface::compile()
 {
 	solAssert(CompilerInputModes.count(m_options.input.mode) == 1);
 	solAssert(!m_assemblyStack);
-	solAssert(!m_compiler);
+	solAssert(!m_evmAssemblyStack && !m_compiler);
 
 	m_compiler = std::make_unique<CompilerStack>(m_universalCallback.callback());
 	m_assemblyStack = m_compiler.get();
@@ -892,8 +935,11 @@ void CommandLineInterface::compile()
 
 void CommandLineInterface::handleCombinedJSON()
 {
-	solAssert(CompilerInputModes.count(m_options.input.mode) == 1);
 	solAssert(m_assemblyStack);
+	solAssert(
+		CompilerInputModes.count(m_options.input.mode) == 1 ||
+		m_options.input.mode == frontend::InputMode::EVMAssemblerJSON
+	);
 
 	if (!m_options.compiler.combinedJsonRequests.has_value())
 		return;